Updated website integration & mask creator
Added a mask creator tool where you can create masks depending on the file, easiest to make masks and parking lot spaces. Added autodetection tool within vision.py so it can tell if there is something in the mask and mark it as occupied or open. FastAPI to the website through all of this.